Module google_indexing3::api

Structs

Indexing
Central instance to access all Indexing related resource activities

PublishUrlNotificationResponse
Output for PublishUrlNotification

UrlNotification
UrlNotification is the resource used in all Indexing API calls. It describes one event in the life cycle of a Web Document.

UrlNotificationGetMetadataCall
Gets metadata about a Web Document. This method can only be used to query URLs that were previously seen in successful Indexing API notifications. Includes the latest UrlNotification received via this API.

UrlNotificationMetadata
Summary of the most recent Indexing API notifications successfully received, for a given URL.

UrlNotificationMethods
A builder providing access to all methods supported on urlNotification resources. It is not used directly, but through the Indexing hub.

UrlNotificationPublishCall
Notifies that a URL has been updated or deleted.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.